Move api/server/router/distribution to daemon/server/router/distribution
